backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
sqlparse==0.5.3
